Start preparing to handle multiple generator types
