+++ generators/csharp/OpenApiCodegen.CSharp.Base/MvcServer/Templates/AddServices.cs
@@ -1,6 +1,6 @@
 ﻿using DarkPatterns.OpenApiCodegen.Handlebars.Templates;
 
-namespace DarkPatterns.OpenApi.CSharp.Templates;
+namespace DarkPatterns.OpenApiCodegen.CSharp.MvcServer.Templates;
 
 public record AddServicesModel(PartialHeader Header, string MethodName, string PackageName, ControllerReference[] Controllers);
