+++ generators/csharp/OpenApiCodegen.CSharp.Base/MvcServer/Templates/Controller.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using DarkPatterns.OpenApiCodegen.Handlebars.Templates;
 
-namespace DarkPatterns.OpenApi.CSharp.Templates;
+namespace DarkPatterns.OpenApiCodegen.CSharp.MvcServer.Templates;
 
 public record ControllerTemplate(
 	PartialHeader Header,
