+++ generators/csharp/OpenApiCodegen.CSharp.Base/MvcServer/Templates/addServices.handlebars
