+++ generators/csharp/OpenApiCodegen.CSharp.Base/MvcServer/Templates/controller.handlebars
